Implemented ability to upload resume files. Needs additional work, but files are uploaded. Also worked on main page application info entry.

+++ JobHuntingApp/JobHuntingApp/Controllers/ResumesController.cs
@@ -3,38 +3,65 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using JobHuntingApp.Data;
 using JobHuntingApp.Models;
+using JobHuntingApp.ViewModels;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Http;
+using Microsoft.Net.Http.Headers;
+using System.IO;
+
+
+
 
 namespace JobHuntingApp.Controllers
 {
+    [Authorize]
     public class ResumesController : Controller
     {
         private readonly JobHuntContext _context;
+        private readonly UserManager<ApplicationUser> _userManager;
+        private IHostingEnvironment _hostingEnvironment;
 
-        public ResumesController(JobHuntContext context)
+        public ResumesController(JobHuntContext context,
+            UserManager<ApplicationUser> userManager, IHostingEnvironment hostingEnvironment)
         {
-            _context = context;    
+            _context = context;
+            _userManager = userManager;
+            _hostingEnvironment = hostingEnvironment;
         }
 
         // GET: Resumes
         public async Task<IActionResult> Index()
         {
-            return View(await _context.Resumes.ToListAsync());
+            var user = await GetCurrentUserAsync();
+            if (user == null)
+            {
+                return View("Error");
+            }
+            return View(await _context.Resumes.Where(r=> r.UserID == user.Id).ToListAsync());
         }
 
         // GET: Resumes/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            var user = await GetCurrentUserAsync();
+            if (user == null)
+            {
+                return View("Error");
+            }
+
             if (id == null)
             {
                 return NotFound();
             }
 
             var resume = await _context.Resumes
-                .SingleOrDefaultAsync(m => m.ResumeID == id);
+                .SingleOrDefaultAsync(m => m.ResumeID == id && m.UserID==user.Id);
             if (resume == null)
             {
                 return NotFound();
@@ -46,34 +73,69 @@ namespace JobHuntingApp.Controllers
         // GET: Resumes/Create
         public IActionResult Create()
         {
-            return View();
+            ResumesViewModel resumesViewModel = new ResumesViewModel();
+            return View(resumesViewModel);
         }
 
         // POST: Resumes/Create
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("ResumeID,UserID,ResumeTitle,ResumeFile")] Resume resume)
+        public async Task<IActionResult> Create(IList<IFormFile> files, string resumeTitle)
         {
-            if (ModelState.IsValid)
+            var user = await GetCurrentUserAsync();
+            if (user == null)
             {
-                _context.Add(resume);
-                await _context.SaveChangesAsync();
-                return RedirectToAction("Index");
+                return View("Error");
             }
-            return View(resume);
+
+            foreach (var file in files)
+            {
+                var fileName = ContentDispositionHeaderValue
+                    .Parse(file.ContentDisposition)
+                    .FileName
+                    .Trim('"');// FileName returns "fileName.ext"(with double quotes) in beta 3
+
+                string filePath = "";
+                if (fileName.EndsWith(".pdf"))// Important for security if saving in webroot
+                {
+                    filePath = _hostingEnvironment.WebRootPath + "\\resumes\\" + fileName;
+                    using (FileStream fs = System.IO.File.Create(filePath))
+                    {
+                        file.CopyTo(fs);
+                        fs.Flush();
+                    }
+                }
+
+                Resume newResume = new Resume()
+                {
+                    UserID = user.Id,
+                    ResumeTitle = resumeTitle,
+                    ResumeFile = filePath
+                };
+
+                _context.Add(newResume);
+                _context.SaveChanges();
+            }
+            return RedirectToAction("Index");// PRG
         }
+
 
         // GET: Resumes/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            var user = await GetCurrentUserAsync();
+            if (user == null)
+            {
+                return View("Error");
+            }
+
             if (id == null)
             {
                 return NotFound();
             }
 
-            var resume = await _context.Resumes.SingleOrDefaultAsync(m => m.ResumeID == id);
+            var resume = await _context.Resumes.SingleOrDefaultAsync(m => m.ResumeID == id && m.UserID==user.Id);
             if (resume == null)
             {
                 return NotFound();
@@ -86,9 +148,15 @@ namespace JobHuntingApp.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ResumeID,UserID,ResumeTitle,ResumeFile")] Resume resume)
+        public async Task<IActionResult> Edit(int id, [Bind("ResumeID,ResumeTitle,UserID,ContentType,ResumeFile")] Resume resume)
         {
-            if (id != resume.ResumeID)
+            var user = await GetCurrentUserAsync();
+            if (user == null)
+            {
+                return View("Error");
+            }
+
+            if (id != resume.ResumeID || resume.UserID != user.Id)
             {
                 return NotFound();
             }
@@ -119,13 +187,19 @@ namespace JobHuntingApp.Controllers
         // GET: Resumes/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            var user = await GetCurrentUserAsync();
+            if (user == null)
+            {
+                return View("Error");
+            }
+
             if (id == null)
             {
                 return NotFound();
             }
 
             var resume = await _context.Resumes
-                .SingleOrDefaultAsync(m => m.ResumeID == id);
+                .SingleOrDefaultAsync(m => m.ResumeID == id && m.UserID == user.Id);
             if (resume == null)
             {
                 return NotFound();
@@ -139,7 +213,14 @@ namespace JobHuntingApp.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var resume = await _context.Resumes.SingleOrDefaultAsync(m => m.ResumeID == id);
+            var user = await GetCurrentUserAsync();
+            if (user == null)
+            {
+                return View("Error");
+            }
+
+
+            var resume = await _context.Resumes.SingleOrDefaultAsync(m => m.ResumeID == id && m.UserID == user.Id);
             _context.Resumes.Remove(resume);
             await _context.SaveChangesAsync();
             return RedirectToAction("Index");
@@ -149,5 +230,12 @@ namespace JobHuntingApp.Controllers
         {
             return _context.Resumes.Any(e => e.ResumeID == id);
         }
+
+        private Task<ApplicationUser> GetCurrentUserAsync()
+        {
+            return _userManager.GetUserAsync(HttpContext.User);
+        }
+
+
     }
 }
